# api/swagger.yaml
openapi: 3.0.1
info:
  title: Data flow 5G Meta API
  description: "APIs exposed to third parties, in order to filter data flows and subscribe\
    \ to the most useful ones."
  version: 1.0.0
servers:
- url: /dataflow-api
...
paths:
...
  /dataflows/{dataType}/query:
    get:
      tags:
      - Dataflow catalogue
      summary: Returns the list of ids of DataFlows that match the query
      operationId: get_data_flows
      parameters:
      - name: dataType
        in: path
        description: Data type of the flow
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: dataSubType
        in: query
        description: Data subType of the flow
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: dataFormat
        in: query
        description: Data format of the flow
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: country
        in: query
        description: Source's country
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: quadkey
        in: query
        description: Source's quadkey
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: sourceId
        in: query
        description: Source's Id
        required: false
        style: form
        explode: true
        schema:
          type: integer
      - name: sourceType
        in: query
        description: Source's type (vehicle or infrastructure)
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: extraParameters
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: object
          additionalProperties:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                example:
                - 1
                - 2
                - 3
                - 5
                - 8
                items:
                  type: integer
                x-content-type: application/json
        "405":
          description: Invalid input
          content: {}
      x-openapi-router-controller: swagger_server.controllers.dataflow_catalogue_controller
  /dataflows/{dataType}/query/count:
    get:
      tags:
      - Dataflow catalogue
      summary: Returns the number of dataFlows that match the query
      operationId: count_data_flows
      parameters:
      - name: dataType
        in: path
        description: Data type of the flow
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: dataSubType
        in: query
        description: Data subType of the flow
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: dataFormat
        in: query
        description: Data format of the flow
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: country
        in: query
        description: Source's country
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: quadkey
        in: query
        description: Source's quadkey
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: sourceId
        in: query
        description: Source's Id
        required: false
        style: form
        explode: true
        schema:
          type: integer
      - name: sourceType
        in: query
        description: Source's type (vehicle or infrastructure)
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: extraParameters
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: object
          additionalProperties:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            text/plain:
              schema:
                type: integer
                example: 5
                x-content-type: text/plain
        "405":
          description: Invalid input
          content: {}
      x-openapi-router-controller: swagger_server.controllers.dataflow_catalogue_controller
...
  /topics:
    get:
      tags:
      - Topic management
      summary: Returns the topics associated with that user
      operationId: topics_by_user
      parameters:
      - name: X_Userinfo
        in: header
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                example:
                - "links_cpm_1, links_cam_1, links_video_1, links_cam_2, links_cam_3"
                items:
                  type: string
                x-content-type: application/json
        "404":
          description: Not found
          content: {}
      x-openapi-router-controller: swagger_server.controllers.topic_management_controller
  /topics/{topicName}:
    get:
      tags:
      - Topic management
      summary: Returns the query that generated the topic
      operationId: find_query_by_topic_name
      parameters:
      - name: topicName
        in: path
        description: Name of the topic
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: X_Userinfo
        in: header
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: successful operation
          content:
            text/plain:
              schema:
                type: string
                x-content-type: text/plain
        "404":
          description: Not found
          content: {}
      x-openapi-router-controller: swagger_server.controllers.topic_management_controller
    delete:
      tags:
      - Topic management
      summary: Delete a topic for the user
      operationId: delete_topic
      parameters:
      - name: topicName
        in: path
        description: Name of the topic to delete
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: X_Userinfo
        in: header
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: successful operation
          content: {}
        "404":
          description: Not found
          content: {}
      x-openapi-router-controller: swagger_server.controllers.topic_management_controller
  /topics/{dataType}/query:
    post:
      tags:
      - Topic management
      summary: catalogue a Kafka topic for the user and returns the topic
      operationId: create_topic
      parameters:
      - name: dataType
        in: path
        description: Data type of the flow
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: dataSubType
        in: query
        description: Data subType of the flow
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: dataFormat
        in: query
        description: Data format of the flow
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: instance_type
        in: query
        description: Instance Type of the pipeline processing the flow
        required: true
        style: form
        explode: true
        schema:
          type: string
      - name: quadkey
        in: query
        description: Source's quadkey
        required: true
        style: form
        explode: true
        schema:
          type: string
      - name: country
        in: query
        description: Source's country
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: sourceId
        in: query
        description: Source's Id
        required: false
        style: form
        explode: true
        schema:
          type: integer
      - name: sourceType
        in: query
        description: Source's type (vehicle or infrastructure)
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: extraParameters
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: object
          additionalProperties:
            type: string
      - name: X_Userinfo
        in: header
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: successful operation
          content:
            text/plain:
              schema:
                type: string
                example: links_cpm_1
                x-content-type: text/plain
        "405":
          description: Invalid input
          content: {}
      x-openapi-router-controller: swagger_server.controllers.topic_management_controller
...